Overlay a bitmap banner on the captured photo
